chore: base adpater use kit v3.0 structure
