+++ packages/iceworks-server/src/lib/adapter-react-v1/index.ts
@@ -2,9 +2,21 @@
  * icescripts 1.0 + kit 2.0
  */
 import baseConfig from '../adapter';
+import Menu from './modules/menu';
+import Router from './modules/router';
+
+const { Configuration, ...resetConfig } = baseConfig;
 
 const config = {
-  ...baseConfig,
+  ...resetConfig,
+  Router: {
+    ...baseConfig.Router,
+    module: Router
+  },
+  Menu: {
+    ...baseConfig.Menu,
+    module: Menu
+  },
 };
 
 export default config;
