+++ packages/iceworks-server/src/lib/adapter-react-v3/README.md
@@ -1,3 +1,3 @@
-# adapter-react-v2
+# adapter-react-v3
 
-适用于 ice-scripts@2.x 的工程和 kit@2.x 的项目最佳实践
+适用于 ice-scripts@2.x 的工程和 kit@3.x 的项目最佳实践
