+++ packages/iceworks-server/src/lib/adapter-react-v3/package.json
@@ -1,3 +1,3 @@
 {
-  "name": "adapter-react-v2"
+  "name": "adapter-react-v3"
 }
